% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{append_metrdata_suffix}
\alias{append_metrdata_suffix}
\title{Append suffix to path}
\usage{
append_metrdata_suffix(path)
}
\arguments{
\item{path}{Path to assign suffix to.}
}
\value{
Path with suffix appended.
}
\description{
Appends the default metadata filename suffix
(\code{"_metadata.json"}) to \code{path}.
}
\examples{
append_metrdata_suffix("~/my_path")
}
\seealso{
\code{\link[=get_metrdata_suffix]{get_metrdata_suffix()}}
}
