% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{cat_metadata}
\alias{cat_metadata}
\title{Concatenate the metadata from to \code{metrdata} objects}
\usage{
cat_metadata(metrdata_obj1, metrdata_obj2)
}
\arguments{
\item{metrdata_obj1}{First \code{metrdata} object to concatenate.}

\item{metrdata_obj2}{Second \code{metrdata} object to concatenate.}
}
\value{
A \code{metrdata} object representing a concatenation of both source
objects.
}
\description{
Concatenate the metadata of the two provide \code{metrdata} objects. Each
object must be a \code{metrdata} object and two objects should not contain
the same keys.
}
\details{
If the two objects contain the same keys, an error or warning is produced
depending on strict mode.

This is an internal helper function. Package users should use the
\code{metrdata()} constructor.
}
