% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{combine_metadata}
\alias{combine_metadata}
\title{Combine metadata to an existing object}
\usage{
combine_metadata(obj, metrdata_obj, prepend = FALSE)
}
\arguments{
\item{obj}{Object or \code{metrdata} object to append metadata to.}

\item{metrdata_obj}{\code{metrdata} object to append to existing object.}

\item{prepend}{If \code{TRUE}, \code{metrdata_obj} is
prepended to head of existing metadata;
otherwise \code{metrdata_obj} is appended to tail of existing metadata.}
}
\value{
\code{obj} with \code{metrdata_obj} appended to its metadata.
}
\description{
This is the internal helper for append/prepend metadata.
}
