% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{convert_to_metrdata}
\alias{convert_to_metrdata}
\title{Convert a list into a \code{metrdata} object}
\usage{
convert_to_metrdata(metadata_list)
}
\arguments{
\item{metadata_list}{List to convert to a \code{metrdata} object.}
}
\value{
\code{metadata_list} as a \code{metrdata} object.
}
\description{
Helper function for setting class of an existing list to a \code{metrdata} object.
Meant for package internal use.
}
