% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{edit_metadata}
\alias{edit_metadata}
\title{Edit object metadata}
\usage{
edit_metadata(
  obj,
  edit_path,
  edit_command = c("<-", "=", "-", "+", "+0", "str+", "str+0", function(cur_value,
    edit_value) {
 }),
  edit_value = NULL,
  ...
)
}
\arguments{
\item{obj}{Object with metadata or \code{metrdata} object to edit.}

\item{edit_path}{Search path for edit, this a vector or list giving the
keys to look up to get to the item to edit, e.g. \verb{c("lvl1a","lvl1a2a"})
will lookup \verb{$lvl1a$lvl1a2a}. A list can be used with a numeric index
to lookup any items by index.}

\item{edit_command}{Type of edit to perform. Supported commands are \verb{<-} or
\code{=} for assignment; \code{-} for deletion; \code{+} for appending (to a list or
vector); \code{+0} for prepending; \verb{str+} for appending to a string;
\code{str+0} for prepending to a string; or an function called with
\verb{(cur_value, edit_value, ...)}.}

\item{edit_value}{Value to assign, append, or prepend. Not used for deletion.}

\item{...}{Additional parameters passed to function given in edit_command.}
}
\value{
A copy of obj with metadata edited.
}
\description{
Edits the metadata of an object or a \code{metrdata} object.
}
\examples{
my_data<-set_metadata(
  data.frame(index=1:10),
  metrdata(
    lvl1a = list(lvl1a2a = "item1", lvl1a2b="item2", lvl1a2c = c(1,2,3)),
    lvl1b = list(lvl1b2a = list(lvl1b2a3a = "xyz"))
  )
)

# edit metadata of object
my_data<-edit_metadata(
  my_data,
  c("lvl1a","lvl1a2d"),"+","item4"
)
metadata(my_data)
}
