% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{get_metadata}
\alias{get_metadata}
\alias{metadata}
\title{Get metadata of an object}
\usage{
get_metadata(obj, as_list = FALSE)

metadata(obj, as_list = FALSE)
}
\arguments{
\item{obj}{Object to obtain metadata from. If \code{obj} is itself a \code{metrdata}
object, than \code{obj} is itself returned.}

\item{as_list}{If \code{TRUE}, simplify return value to a list, instead of
returning a \code{metrdata} object.}
}
\value{
\code{metrdata} object or list containing the metadata of \code{obj}.
}
\description{
These functions return the metadata of the provided object. An error is
produced if the object does not have any associated metadata.
}
\details{
The use of \code{get_metadata()} is preferred as the intention is more explicit;
\code{metadata()} is provided as an alias for completeness.
}
\examples{
a_data_frame <- data.frame(index = 1:10)
metadata(a_data_frame) <- metrdata(list(note = "Indexes 1 to 10"))
get_metadata(a_data_frame)
metadata(a_data_frame)
}
