% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{get_metrdata_attrname}
\alias{get_metrdata_attrname}
\title{Get the attribute name used for storing metadata associated with objects}
\usage{
get_metrdata_attrname()
}
\value{
Returns \code{"metrdata"}.
}
\description{
Returns the attribute name used for storing \code{metrdata} objects that are
associated with another object as metadata (which is \code{metrdata}).
}
\examples{
get_metrdata_attrname()
}
