% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{get_metrdata_classname}
\alias{get_metrdata_classname}
\title{Get the class name used for \code{metrdata} objects}
\usage{
get_metrdata_classname()
}
\value{
Returns \code{"metrdata"}.
}
\description{
Returns the class name used for \code{metrdata} objects (which is
\code{metrdata}).
}
\examples{
get_metrdata_classname()
}
