% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{get_metrdata_suffix}
\alias{get_metrdata_suffix}
\title{Get the filename suffix used for writing metadata sidecar files}
\usage{
get_metrdata_suffix()
}
\value{
Returns \code{"_metadata.json"}.
}
\description{
Returns the filename suffix used for writing metadata sidecar files
(\verb{_metadata.json}).
}
\examples{
get_metrdata_suffix()
}
\seealso{
\code{\link[=append_metrdata_suffix]{append_metrdata_suffix()}}
}
