% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{has_is_metrdata}
\alias{has_is_metrdata}
\title{Test whether object has metadata or is a \code{metrdata} object}
\usage{
has_is_metrdata(this)
}
\arguments{
\item{this}{Object to test.}
}
\value{
\code{TRUE} if \code{this} is a \code{metrdata} object or has metadata,
otherwise \code{FALSE}.
}
\description{
This function tests if an object is itself a \code{metrdata} object, or
has any associated metadata.
}
\examples{
obj_metrdata <- metrdata(list(note = "Indexes 1 to 10"))
has_is_metrdata(obj_metrdata) # TRUE
is_metrdata(list()) # FALSE
a_data_frame <- data.frame(index = 1:10)
has_is_metrdata(a_data_frame) # FALSE
metadata(a_data_frame) <- obj_metrdata
has_is_metrdata(a_data_frame) # TRUE
}
\seealso{
\code{\link[=is_metrdata]{is_metrdata()}} \code{\link[=has_metadata]{has_metadata()}}
}
