% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{has_metadata}
\alias{has_metadata}
\title{Check whether object has metadata}
\usage{
has_metadata(obj)
}
\arguments{
\item{obj}{Object to test for presence of metadata.}
}
\value{
\code{TRUE} if object has metadata, otherwise \code{FALSE}
}
\description{
This function checks whether the provided object has any associated metadata.
}
\examples{
a_data_frame <- data.frame(index = 1:10)
has_metadata(a_data_frame) # FALSE
metadata(a_data_frame) <- metrdata(list(note = "Indexes 1 to 10"))
has_metadata(a_data_frame) # TRUE
stripped_data_frame <- strip_metadata(a_data_frame)
has_metadata(stripped_data_frame) # FALSE
}
