% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{include_metrdata}
\alias{include_metrdata}
\title{Process \code{metrdata} 'includes'}
\usage{
include_metrdata(include_from)
}
\arguments{
\item{include_from}{List of objects with metadata or \code{metrdata} objects to
include.}
}
\value{
A new \code{metrdata} list with contents of all objects in \code{include_from}
concatenated.
}
\description{
This is a helper function for processing other \code{metrdata} objects 'included'
when constructing a new \code{metrdata} object.
}
\section{Includes}{


An 'include' is an object with metadata or a \code{metrdata} object whose
contents is included directly in the new object, i.e. their contents are
flattened into the new object.
}

