% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{is_metrdata}
\alias{is_metrdata}
\title{Test whether object is a \code{metrdata} object}
\usage{
is_metrdata(this)
}
\arguments{
\item{this}{Object to test.}
}
\value{
\code{TRUE} if \code{this} is a \code{metrdata} object, otherwise \code{FALSE}.
}
\description{
This function tests if an object is a \code{metrdata} object.
}
\examples{
is_metrdata(metrdata()) # TRUE
is_metrdata(list()) # FALSE
}
