% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{prepend_metadata}
\alias{prepend_metadata}
\alias{append_metadata}
\title{Append/prepend metadata to an existing object}
\usage{
prepend_metadata(obj, metrdata_obj)

append_metadata(obj, metrdata_obj)
}
\arguments{
\item{obj}{Object or \code{metrdata} object to append metadata to.}

\item{metrdata_obj}{\code{metrdata} object to append to existing object.}
}
\description{
Append or prepend additional metadata to an existing object with
metadata or \code{metrdata} object. If the the existing object does not already
have metadata, the provided metadata is used directly as the object's
metadata.
}
\examples{
append_metadata(
  set_metadata(data.frame(index = 1:10),
    metrdata(`Original data source` = list(Note="Fictional data", Created="2024-01"))
    ),
  metrdata(Note="This is an included metadata"))
prepend_metadata(
  set_metadata(data.frame(index = 1:10),
    metrdata(`Original data source` = list(Note="Fictional data", Created="2024-01"))
    ),
  metrdata(Note="This is an included metadata"))
}
