% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{preprocess_metadata}
\alias{preprocess_metadata}
\title{Preprocess a list into metadata}
\usage{
preprocess_metadata(metadata_list)
}
\arguments{
\item{metadata_list}{A \code{list} to process into a \code{metrdata} object.}
}
\value{
A \code{metrdata} object processed from \code{metadata_list}.
}
\description{
Preprocess a list into a \code{metrdata} object.
This function is not meant to be called directly; it is an internal helper
meant to be called from the \code{metrdata} constructor.
}
\section{Inclusion of nested lists}{


Child lists are processed when converting to a \code{metrdata} object by
parsing its contents and directly embedding metadata from any children that
themselves are \code{metrdata} objects or have associated metadata.

This allows a pattern where an object containing metadata may be referenced
when creating metadata for an object.
}

