% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata from a sidecar file}
\usage{
read_metadata(path, add_suffix = TRUE, ...)
}
\arguments{
\item{path}{Path to read metadata from.}

\item{add_suffix}{If \code{TRUE}, the standard suffix for metadata sidecar files
is automatically appended to \code{path} before reading.}

\item{...}{Additional parameters passed to \code{jsonlite::read_json} for reading.}
}
\value{
Metadata as a \code{metrdata} object.
}
\description{
Metadata in this package is stored in sidecar files in JSON format alongside
the data file for which the metadata is associated with. The sidecar file
should have the same name as the data file but with the suffix
"_metadata.json appended.
This function reads metadata (only) from a sidecar file.
}
\examples{
a_data_frame <- data.frame(index = 1:10)
metadata(a_data_frame) <- metrdata(list(note = "Indexes 1 to 10"))
tmpfilename<-tempfile()
write_metadata(a_data_frame,tmpfilename) # saves to tmpfilename + suffix
reread_metadata<-read_metadata(tmpfilename)
testthat::expect_equal(get_metadata(a_data_frame), reread_metadata)
file.remove(metrdata:::append_metrdata_suffix(tmpfilename))
}
\seealso{
\code{\link[=write_metadata]{write_metadata()}}
}
