% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{set_strict}
\alias{set_strict}
\title{Set strict behaviour}
\usage{
set_strict(strict_status = TRUE)
}
\arguments{
\item{strict_status}{Status to set to.}
}
\value{
None.
}
\description{
This function enables or disables strict behaviour in this package.
Strict behaviour is enabled by default and causes errors for
non-critical issues when processing metadata, instead of just warnings.
}
\examples{
set_strict(FALSE)
metrdata:::stop_or_warn("problem encountered") # warning only
set_strict(TRUE)
}
