% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{strip_metadata}
\alias{strip_metadata}
\title{Remove metadata from an object}
\usage{
strip_metadata(obj)
}
\arguments{
\item{obj}{Object to strip metadata from.}
}
\value{
Object without metadata.
}
\description{
This function removes metadata from the provided object.
An error/warning is produced if the object does not have any associated
metadata.
}
\examples{
a_data_frame <- data.frame(index = 1:10)
metadata(a_data_frame) <- metrdata(list(note = "Indexes 1 to 10"))
has_metadata(a_data_frame) # TRUE
stripped_data_frame <- strip_metadata(a_data_frame)
has_metadata(stripped_data_frame) # FALSE
}
