% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{write_metadata}
\alias{write_metadata}
\title{Write metadata to a sidecar file}
\usage{
write_metadata(obj, path, pretty = 2, add_suffix = TRUE, ...)
}
\arguments{
\item{obj}{Object containing metadata or \code{metrdata} object to write.}

\item{path}{Path to write metadata to.}

\item{pretty}{adds indentation whitespace to JSON output. Can be TRUE/FALSE or a number specifying the number of spaces to indent. See \code{\link[jsonlite:prettify]{prettify()}}}

\item{add_suffix}{If \code{TRUE}, the standard suffix for metadata sidecar files
is automatically appended to \code{path} before reading.}

\item{...}{Additional parameters based to \code{jsonlite::write_json()} for writing.}
}
\value{
None.
}
\description{
This function writes metadata (only) to a sidecar file. It can be used to
write metadata associated with an object, or a \code{metrdata} object directly.
}
\examples{
a_data_frame <- data.frame(index = 1:10)
metadata(a_data_frame) <- metrdata(list(note = "Indexes 1 to 10"))
tmpfilename<-tempfile()
write_metadata(a_data_frame,tmpfilename) # saves to tmpfilename + suffix
reread_metadata<-read_metadata(tmpfilename)
testthat::expect_equal(get_metadata(a_data_frame), reread_metadata)
file.remove(metrdata:::append_metrdata_suffix(tmpfilename))
}
\seealso{
\code{\link[=read_metadata]{read_metadata()}}
}
