% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrdata.R
\name{write_with_metadata}
\alias{write_with_metadata}
\alias{write_tsv_metadata}
\alias{write_csv_metadata}
\alias{write_csv2_metadata}
\alias{write_delim_metadata}
\title{Write a file together with its metadata}
\usage{
write_with_metadata(
  obj,
  path,
  write_func,
  ...,
  metadata_args = list(pretty = 2)
)

write_tsv_metadata(obj, path, ..., metadata_args = list(pretty = 2))

write_csv_metadata(obj, path, ..., metadata_args = list(pretty = 2))

write_csv2_metadata(obj, path, ..., metadata_args = list(pretty = 2))

write_delim_metadata(obj, path, ..., metadata_args = list(pretty = 2))
}
\arguments{
\item{obj}{Object to write to file. The object must contain associated
metadata.}

\item{path}{Path to write data to The path for the metadata sidecar file
is automatically derived by appending the expected suffix to this path.}

\item{write_func}{For \code{write_with_metadata}, the user should provide the
function used for writing the data to \code{path}.}

\item{...}{Additional parameters passed to the write function.}

\item{metadata_args}{Additional parameters passed to \code{write_metadata()}.}
}
\value{
None.
}
\description{
This series of functions writes an object to file together with its
associated metadata.
The metadata is written in a JSON sidecar file alongside the data file.
}
\details{
\code{write_with_metadata()} allows a data file to be written with a user-provided
function for writing to \code{path}.

\code{write_tsv_metadata()} writes the data file to \code{path} using \code{readr::write_tsv()}.

\code{write_csv_metadata()} writes the data file to \code{path} using \code{readr::write_csv()}.

\code{write_csv2_metadata()} writes the data file to \code{path} using \code{readr::write_csv2()}.

\code{write_delim_metadata()} writes the data file to \code{path} using \code{readr::write_delim()}.
}
\seealso{
\code{\link[=write_metadata]{write_metadata()}} \code{\link[readr:write_delim]{readr::write_tsv()}} \code{\link[readr:write_delim]{readr::write_csv()}}
\code{\link[readr:write_delim]{readr::write_csv2()}} \code{\link[readr:write_delim]{readr::write_delim()}}
}
